requirements.txt
numpy
matplotlib
pillow
pycryptodome
